# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Insurance Reasoning Engine (IRE)",
    "description": "Insurance Reasoning Engine (IRE) - Clean Architecture FastAPI Backend Skeleton",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/ocr/extract": {
      "post": {
        "tags": [
          "Modular OCR Service"
        ],
        "summary": "Extract Document Ocr",
        "description": "Synchronous Document OCR Extraction Endpoint.\nExtracts text, handwriting, language, layout regions, bounding boxes, tables, and confidence scores into standardized JSON.",
        "operationId": "extract_document_ocr_api_v1_ocr_extract_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_extract_document_ocr_api_v1_ocr_extract_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ocr/async_extract": {
      "post": {
        "tags": [
          "Modular OCR Service"
        ],
        "summary": "Async Extract Document Ocr",
        "description": "Asynchronous Document OCR Job Submission Endpoint.\nEnqueues OCR job in background queue with exponential backoff & retries.",
        "operationId": "async_extract_document_ocr_api_v1_ocr_async_extract_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_async_extract_document_ocr_api_v1_ocr_async_extract_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ocr/tasks/{task_id}": {
      "get": {
        "tags": [
          "Modular OCR Service"
        ],
        "summary": "Get Ocr Task Status",
        "description": "Poll Async OCR Task Status and Results.",
        "operationId": "get_ocr_task_status_api_v1_ocr_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_async_extract_document_ocr_api_v1_ocr_async_extract_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "document_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_async_extract_document_ocr_api_v1_ocr_async_extract_post"
      },
      "Body_extract_document_ocr_api_v1_ocr_extract_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_extract_document_ocr_api_v1_ocr_extract_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/hospitals/login"
          }
        }
      }
    }
  }
}